Update in comments; fix unit test for notification on workspace setup.

+++ src/lib/sequencer-gui-components/sequencergui/pvmonitor/sequencer_workspace_listener.cpp
@@ -97,11 +97,8 @@ void SequencerWorkspaceListener::StopListening()
 
 int SequencerWorkspaceListener::GetEventCount() const
 {
-  return p_impl->m_workspace_events.size();
+  return static_cast<int>(p_impl->m_workspace_events.size());
 }
-
-//! Takes a workspace event from the queue and returns it to the user.
-//! Intended for call from the GUI thread.
 
 WorkspaceEvent SequencerWorkspaceListener::PopEvent() const
 {
